requirements.txt
deepgram_sdk==3.3.1
langchain==0.2.9
langchain_core==0.2.21
langchain_groq==0.1.6
playsound==1.3.0
pygobject